demo/requirements.txt
great-expectations==1.4.4
pandas==2.1.4
papermill==2.6.*
ipykernel==6.29.5
apache-airflow-providers-papermill==3.11.0
gdown==5.2.*
minio==7.2.15
apache-airflow-providers-amazon==9.8.*
matplotlib==3.10.3
scipy==1.15.3
scikit-learn==1.6.1
seaborn==0.13.2
geopandas==1.0.1
shapely==2.1.1